+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Unreleased
 ### Changed
+* Pyarrow >=17 required
 * Partitioning supports original indices
 * Acero engine declaration
 * `Duration` format improvements